+++ openstack/autoscaling/v1/groups/enable.go
@@ -1,0 +1,11 @@
+package groups
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+// Enable is an operation by which can make the group enable service
+func Enable(client *golangsdk.ServiceClient, id string) (r ActionResult) {
+	opts := ActionOpts{
+		Action: "resume",
+	}
+	return doAction(client, id, opts)
+}
